+++ lib/cli/tests/init.test.ts
