+++ templates/minimal/tests/test.ts
@@ -1,5 +1,5 @@
 import { fetchHtml } from "$fresh/tests/test_utils.ts";
-import { withFresh } from "./test_utils.ts";
+import { withFresh } from "./test.utils.ts";
 import { assertEquals } from "std/assert/mod.ts";
 import { manifest } from "$fresh/src/dev/mod.ts";
